Steps/02-StreamHot/changestream/python/requirements.txt
pymongo==3.10.0
azure-eventhub